Updated to sveltekit-next
