+++ .gitignore
@@ -6,5 +6,4 @@ node_modules
 .env
 .env.*
 !.env.example
-static/cache
-.pnpm-debug.log
+OOLD
