--- .storybook/manager.js
@@ -1,24 +1,0 @@
-import { addons } from '@storybook/addons';
-
-addons.setConfig({
-    isFullscreen: false,
-    showNav: true,
-    showPanel: true,
-    panelPosition: 'right',
-    enableShortcuts: true,
-    isToolshown: true,
-    theme: undefined,
-    selectedPanel: undefined,
-    initialActive: 'sidebar',
-    sidebar: {
-      showRoots: false,
-      collapsedRoots: ['other'],
-    },
-    toolbar: {
-      title: { hidden: false, },
-      zoom: { hidden: false, },
-      eject: { hidden: false, },
-      copy: { hidden: false, },
-      fullscreen: { hidden: false, },
-    },
-  });
