--- .storybook/preview-head.html
@@ -1,2 +1,0 @@
-<!-- .storybook/preview-head.html -->
-<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
