--- .vscode/launch.json
@@ -1,18 +1,0 @@
-{
-    "configurations": [
-    {
-        "name": "Launch via NPM",
-        "request": "launch",
-        "runtimeArgs": [
-            "run",
-            "dev"
-        ],
-        "runtimeExecutable": "npm",
-        "skipFiles": [
-            "<node_internals>/**"
-        ],
-        "type": "pwa-node"
-    }
-
-    ]
-}
