--- .vscode/settings.json
@@ -1,21 +1,0 @@
-{
-    "files.exclude": {
-        "**/.git": false,
-        "**/.svn": false,
-        "**/.hg": false,
-        "**/CVS": false,
-        "**/.DS_Store": false,
-        "**/Thumbs.db": false,
-        "**/.classpath": false,
-        "**/.project": false,
-        "**/.settings": false,
-        "**/.factorypath": false,
-        "node_modules": false,
-        "**/.svelte-kit": false,
-        ".svelte-kit": false,
-        "build": false,
-        ".nvmrc": false,
-        ".vscode": false
-    },
-    "explorerExclude.backup": null
-}
