+++ README.md
@@ -8,13 +8,11 @@ If you're seeing this, you've probably already done this step. Congrats!
 
 ```bash
 # create a new project in the current directory
-npm init svelte@next
+npm create svelte@latest
 
 # create a new project in my-app
-npm init svelte@next my-app
+npm create svelte@latest my-app
 ```
-
-> Note: the `@next` is temporary
 
 ## Developing
