+++ package.json
@@ -1,70 +1,48 @@
 {
-  "name": "tvheadend-guide",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "svelte-kit dev",
-    "clean": "rm -rf ./build",
-    "build": "svelte-kit build",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
-    "test": "playwright test",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint --ignore-path .gitignore .",
-    "format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. .",
-    "tsc": "npx tsc --outdir build",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@playwright/test": "^1.19.1",
-    "@storybook/addon-actions": "^6.4.22",
-    "@storybook/addon-essentials": "^6.4.22",
-    "@storybook/addon-links": "^6.4.22",
-    "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/addon-storysource": "^6.4.22",
-    "@storybook/addon-svelte-csf": "^2.0.2",
-    "@storybook/addons": "^6.4.22",
-    "@storybook/api": "^6.4.22",
-    "@storybook/client-api": "^6.4.22",
-    "@storybook/client-logger": "^6.4.22",
-    "@storybook/components": "^6.4.22",
-    "@storybook/core-events": "^6.4.22",
-    "@storybook/svelte": "^6.4.22",
-    "@storybook/theming": "^6.4.22",
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/adapter-node": "^1.0.0-next.71",
-    "@sveltejs/kit": "next",
-    "@tailwindcss/forms": "^0.5.0",
-    "@typescript-eslint/eslint-plugin": "^5.10.1",
-    "@typescript-eslint/parser": "^5.10.1",
-    "autoprefixer": "^10.4.2",
-    "babel-loader": "^8.2.4",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-storybook": "^0.5.10",
-    "eslint-plugin-svelte3": "^3.2.1",
-    "postcss": "^8.4.7",
-    "prettier": "^2.5.1",
-    "prettier-plugin-svelte": "^2.5.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "svelte": "^3.47.0",
-    "svelte-check": "^2.2.6",
-    "svelte-loader": "^3.1.2",
-    "svelte-preprocess": "^4.10.1",
-    "tailwindcss": "^3.0.23",
-    "tslib": "^2.3.1",
-    "typescript": "~4.5.4",
-    "webpack": "^5.72.0"
-  },
-  "type": "module",
-  "dependencies": {
-    "anylogger": "^1.0.11",
-    "anylogger-console": "^1.0.0",
-    "axios": "^0.26.0",
-    "dateformat": "^5.0.3",
-    "uuid": "^8.3.2"
-  }
+	"name": "tvheadend-guide",
+	"version": "0.0.1",
+	"scripts": {
+		"dev": "vite dev",
+		"build": "vite build",
+		"package": "svelte-kit package",
+		"preview": "vite preview",
+		"prepare": "svelte-kit sync",
+		"test": "playwright test",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
+		"format": "prettier --write --plugin-search-dir=. ."
+	},
+	"devDependencies": {
+		"@playwright/test": "^1.22.2",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
+		"@tailwindcss/forms": "^0.5.2",
+		"@typescript-eslint/eslint-plugin": "^5.27.0",
+		"@typescript-eslint/parser": "^5.27.0",
+		"autoprefixer": "^10.4.7",
+		"eslint": "^8.16.0",
+		"eslint-config-eslint": "^5.0.1",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-node": "^6.0.1",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"postcss": "^8.4.14",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"svelte": "^3.44.0",
+		"svelte-check": "^2.7.1",
+		"svelte-preprocess": "^4.10.7",
+		"tailwindcss": "^3.1.6",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.4",
+		"vite": "^3.0.0"
+	},
+	"type": "module",
+	"dependencies": {
+		"anylogger": "^1.0.11",
+		"anylogger-console": "^1.0.0",
+		"axios": "^0.27.2",
+		"dateformat": "^5.0.3",
+		"uuid": "^8.3.2"
+	}
 }
