--- playwright.config.js
@@ -1,9 +1,0 @@
-/** @type {import('@playwright/test').PlaywrightTestConfig} */
-const config = {
-	webServer: {
-		command: 'npm run build && npm run preview',
-		port: 3000
-	}
-};
-
-export default config;
