+++ src/lib/components/NavRoute.ts
@@ -1,0 +1,9 @@
+export class NavRoute{
+		constructor(
+			public path: string,
+			public segment: string,
+			public label: string,
+			public icon: string|undefined = undefined
+		) {
+		}
+    }
