+++ src/routes/api/v1/epg/channeltags/index.ts
@@ -3,7 +3,7 @@ import type { RequestHandlerOutput } from "@sveltejs/kit";
 import {tvhCache} from "$lib/server/tvh/tvh-cache"
 
 /** @type {import('./').RequestHandler} */
-export async function get() :Promise<RequestHandlerOutput>{
+export async function GET() :Promise<RequestHandlerOutput>{
 
     const tags = tvhCache.channelTags
     const body = {}
