+++ src/routes/api/v1/epg/events/[id].ts
@@ -5,7 +5,7 @@ import {httpErr404} from "$lib/server/ApiHelper"
 import {tvhCache} from "$lib/server/tvh/tvh-cache"
 
 /** @type {import('./').RequestHandler} */
-export async function get({params}) :Promise<RequestHandlerOutput>{
+export async function GET({params}) :Promise<RequestHandlerOutput>{
     const epgid = params.id;
     if (!tvhCache.epg.has(epgid)) {
         return httpErr404("EPG Event not found",params)
