+++ src/routes/api/v1/epg/events/index.ts
@@ -5,7 +5,7 @@ import {epgEventsQuery, EPGFilter} from "$lib/server/ApiHelper"
 import {tvhCache} from "$lib/server/tvh/tvh-cache"
 
 /** @type {import('./').RequestHandler} */
-export async function get({url}) :Promise<RequestHandlerOutput>{
+export async function GET({url}) :Promise<RequestHandlerOutput>{
   const body = {}
   const filter= new EPGFilter(tvhCache);
   epgEventsQuery(filter, url, body,tvhCache.epgSorted);
