+++ src/routes/api/v1/health.ts
@@ -1,7 +1,7 @@
 import type { RequestHandlerOutput } from "@sveltejs/kit";
 
 /** @type {import('./').RequestHandler} */
-export async function get() :Promise<RequestHandlerOutput>{
+export async function GET() :Promise<RequestHandlerOutput>{
 
       return {
         body: { serviceHealth:"OK",
