--- src/stories/examples/Header.stories.svelte
@@ -1,36 +1,0 @@
-<script>
-  import { Meta, Template, Story } from "@storybook/addon-svelte-csf";
-  import Header from "./Header.svelte";
-</script>
-
-<Meta
-  title="Example/Header"
-  component={Header}
-  argTypes={{
-    onLogin: { action: "onLogin" },
-    onLogout: { action: "onLogout" },
-    onCreateAccount: { action: "onCreateAccount" },
-  }}
-/>
-
-<Template let:args>
-  <Header
-    {...args}
-    on:login={args.onLogin}
-    on:logout={args.onLogout}
-    on:createAccount={args.onCreateAccount}
-  />
-  <br>
-  {#if args.user}
-  {args.user.name}
-  {/if}
-</Template>
-
-<Story
-  name="LoggedIn"
-  args={{
-    user: {},
-  }}
-/>
-
-<Story name="LoggedOut" args={{}} />
