+++ svelte.config.js
@@ -1,24 +1,14 @@
-import adapter from '@sveltejs/adapter-node';
+import adapter from '@sveltejs/adapter-auto';
 import preprocess from 'svelte-preprocess';
-
-import path from "path"
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	// Consult https://github.com/sveltejs/svelte-preprocess
 	// for more information about preprocessors
-	vite: {
-		resolve: {
-		  alias: {
-			$lib: path.resolve('./src/lib'),
-			$components: path.resolve('./src/lib/components'),
-		  },
-		},
-	  },
 	preprocess: preprocess(),
 
 	kit: {
-		adapter: adapter({ out: 'build' })
+		adapter: adapter()
 	}
 };
