--- thunder-collection_local.json
@@ -1,1 +1,0 @@
-{"client":"Thunder Client","collectionName":"local","dateExported":"2022-03-09T15:15:07.072Z","version":"1.1","folders":[],"requests":[{"containerId":"","sortNum":10000,"headers":[{"name":"Accept","value":"*/*"},{"name":"User-Agent","value":"Thunder Client (https://www.thunderclient.com)"}],"colId":"f091245b-23bb-4e3f-a961-db97035fdad6","name":"http://localhost:3000/api/sandbox/simple-tvh","url":"http://localhost:3000/api/sandbox/simple-tvh","method":"GET","modified":"2022-03-09T15:11:47.603Z","created":"2022-03-09T14:57:58.835Z","_id":"d6813f96-595f-464d-b0f9-62cb4b4d1e0c","params":[],"tests":[]}]}
